+++ vector/src/main/java/im/vector/app/features/home/avatar/DefaultAvatarStyle.kt
@@ -15,10 +15,11 @@ enum class DefaultAvatarStyle(@StringRes val titleRes: Int) {
     ELEMENT(CommonStrings.avatar_style_element),
     GENERIC(CommonStrings.avatar_style_generic),
     TWITTER_EGG(CommonStrings.avatar_style_twitter_egg),
+    MATRIX(CommonStrings.avatar_style_matrix),
     HASHTAG(CommonStrings.avatar_style_hashtag);
 
     companion object {
-        val PEOPLE = listOf(ELEMENT, GENERIC, TWITTER_EGG)
+        val PEOPLE = listOf(ELEMENT, GENERIC, TWITTER_EGG, MATRIX)
         val ROOM = listOf(ELEMENT, HASHTAG)
 
         fun of(name: String?, among: List<DefaultAvatarStyle>) = among.firstOrNull { it.name == name } ?: ELEMENT